update crypto-simd portable impl to trait-simd API
- SIMD support crates: reorganize to reduce unneeded abstraction. One
  crate with platform-independent and platform-specific components.
- Chacha: more separation between optional frontends.
- Chacha / JH / BLAKE: update for new SIMD crate structure; remove
  references to packed_simd (currently unsupported as backend) / ppv_null
  (to keep thing simple, SIMD is now mandatory on supported platforms).
- Chacha: fix endian-portability
- Update portable fallback for new API
