+++ hashes/blake/Cargo.toml
@@ -10,17 +10,13 @@ categories = ["cryptography", "no-std"]
 
 [dependencies]
 block-buffer = "0.7"
-crypto-simd = { version = "0.2" }
 digest = "0.8"
 lazy_static = { version = "1.2", optional = true }
-packed_simd_crate = { package = "packed_simd", version = "0.3.2", optional = true }
-ppv-null = { version = "0.2" }
 simd = { package = "ppv-lite86", version = "0.2.0", optional = true }
 
 [features]
 default = ["simd", "std"]
 std = ["lazy_static"]
-packed_simd = ["packed_simd_crate", "crypto-simd/packed_simd"]
 
 [dev-dependencies]
 digest = { version = "0.8", features = ["dev"] }
