+++ hashes/jh/Cargo.toml
@@ -12,13 +12,10 @@ edition = "2018"
 
 [dependencies]
 block-buffer = "0.7"
-crypto-simd = { version = "0.2" }
 digest = "0.8"
 hex-literal = "0.1"
 lazy_static = { version = "1.2", optional = true }
-packed_simd_crate = { package = "packed_simd", version = "0.3.2", optional = true }
-ppv-null = { version = "0.2" }
-simd = { package = "ppv-lite86", version = "0.2", optional = true }
+simd = { package = "ppv-lite86", version = "0.2" }
 
 [dev-dependencies]
 digest = { version = "0.8", features = ["dev"] }
@@ -27,10 +24,8 @@ digest = { version = "0.8", features = ["dev"] }
 cc = "1.0.3"
 
 [features]
-default = ["std","simd"]
+default = ["std"]
 std = ["lazy_static"]
-avx2 = ["simd"]
-packed_simd = ["packed_simd_crate", "crypto-simd/packed_simd"]
 
 [badges]
 travis-ci = { repository = "cryptocorrosion/cryptocorrosion" }
