+++ stream-ciphers/chacha/Cargo.toml
@@ -13,22 +13,18 @@ documentation = "https://docs.rs/c2-chacha"
 
 [dependencies]
 byteorder = "1.3"
-crypto-simd = "0.2.0"
 generic-array = "0.13.0"
 lazy_static = { version = "1.2", optional = true }
-packed_simd_crate = { package = "packed_simd", version = "0.3.2", optional = true }
-ppv-null = "0.2.0"
 rand_core = { version = "0.4.0", optional = true }
-simd = { package = "ppv-lite86", version = "0.2.0", optional = true }
+simd = { package = "ppv-lite86", version = "0.2.0" }
 stream-cipher = { version = "0.3", optional = true }
 
 [dev-dependencies]
 hex-literal = "0.1"
 
 [features]
-default = ["simd", "std", "rustcrypto_api"]
+default = ["std", "rustcrypto_api", "provide_rand_backend"]
 std = ["lazy_static"]
-packed_simd = ["packed_simd_crate", "crypto-simd/packed_simd"]
 rustcrypto_api = ["stream-cipher"]
 provide_rand_backend = ["rand_core"]
