+++ stream-ciphers/chacha/src/lib.rs
@@ -2,7 +2,7 @@
 
 //! Pure Rust ChaCha with SIMD optimizations.
 //!
-//! Usage:
+//! Stream-cipher usage:
 //! ```
 //! extern crate c2_chacha;
 //!
@@ -30,1195 +30,25 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-extern crate byteorder;
-extern crate crypto_simd;
 #[cfg(test)]
 #[macro_use]
 extern crate hex_literal;
+
 #[cfg(feature = "std")]
 #[macro_use]
 extern crate lazy_static;
-#[cfg(feature = "rustcrypto_api")]
-pub extern crate stream_cipher;
 
-#[cfg(feature = "packed_simd")]
-extern crate packed_simd_crate;
-#[cfg(not(any(feature = "simd", feature = "packed_simd")))]
-extern crate ppv_null;
-#[cfg(all(feature = "simd", not(feature = "packed_simd")))]
 #[macro_use]
-pub extern crate simd;
-#[cfg(feature = "packed_simd")]
-use crypto_simd::packed_simd::u32x4x4;
-#[cfg(feature = "packed_simd")]
-use packed_simd_crate::u32x4;
-#[cfg(not(any(feature = "simd", feature = "packed_simd")))]
-use ppv_null::{u32x4, u32x4x4};
-#[cfg(all(feature = "simd", not(feature = "packed_simd")))]
-use simd::Machine;
+extern crate simd;
 
-use byteorder::{ByteOrder, LE};
-use core::{cmp, u32, u64};
-
-#[cfg(not(feature = "rustcrypto_api"))]
-extern crate generic_array;
-#[cfg(feature = "rustcrypto_api")]
-use stream_cipher::generic_array;
-
-use self::generic_array::typenum::{Unsigned, U10, U12, U24, U32, U4, U6, U8};
-use self::generic_array::{ArrayLength, GenericArray};
-
-use simd::{vec128_storage, ArithOps, BitOps32, LaneWords4, MultiLane, StoreBytes, Vec2, Vec4};
-
-const BLOCK: usize = 64;
-const BLOCK64: u64 = BLOCK as u64;
-const LOG2_BUFBLOCKS: u64 = 2;
-const BUFBLOCKS: u64 = 1 << LOG2_BUFBLOCKS;
-const BUFSZ64: u64 = BLOCK64 * BUFBLOCKS;
-const BUFSZ: usize = BUFSZ64 as usize;
-
-const BIG_LEN: u64 = 0;
-const SMALL_LEN: u64 = 1 << 32;
-
-#[derive(Clone)]
-pub struct State<V> {
-    pub a: V,
-    pub b: V,
-    pub c: V,
-    pub d: V,
-}
-
-#[inline(always)]
-pub fn round<V: ArithOps + BitOps32>(mut x: State<V>) -> State<V> {
-    x.a += x.b;
-    x.d = (x.d ^ x.a).rotate_each_word_right16();
-    x.c += x.d;
-    x.b = (x.b ^ x.c).rotate_each_word_right20();
-    x.a += x.b;
-    x.d = (x.d ^ x.a).rotate_each_word_right24();
-    x.c += x.d;
-    x.b = (x.b ^ x.c).rotate_each_word_right25();
-    x
-}
-
-#[inline(always)]
-pub fn diagonalize<V: LaneWords4>(mut x: State<V>) -> State<V> {
-    x.b = x.b.shuffle_lane_words3012();
-    x.c = x.c.shuffle_lane_words2301();
-    x.d = x.d.shuffle_lane_words1230();
-    x
-}
-#[inline(always)]
-pub fn undiagonalize<V: LaneWords4>(mut x: State<V>) -> State<V> {
-    x.b = x.b.shuffle_lane_words1230();
-    x.c = x.c.shuffle_lane_words2301();
-    x.d = x.d.shuffle_lane_words3012();
-    x
-}
-
-impl ChaCha {
-    /// Set 64-bit block count, affecting next refill.
-    #[inline(always)]
-    fn seek64<M: Machine>(&mut self, m: M, blockct: u64) {
-        // x86 is little-endian
-        let d: M::u64x2 = m.unpack(self.d);
-        self.d = d.insert(blockct, 0).into();
-    }
-
-    /// Set 32-bit block count, affecting next refill.
-    #[inline(always)]
-    fn seek32<M: Machine>(&mut self, m: M, blockct: u32) {
-        let d: M::u32x4 = m.unpack(self.d);
-        self.d = d.insert(blockct, 0).into();
-    }
-
-    /// Produce output from the current state.
-    #[inline(always)]
-    fn output_narrow<M: Machine>(&mut self, m: M, x: State<M::u32x4>, out: &mut [u8; BLOCK]) {
-        let k = m.vec([0x6170_7865, 0x3320_646e, 0x7962_2d32, 0x6b20_6574]);
-        (x.a + k).write_le(&mut out[0..16]);
-        (x.b + m.unpack(self.b)).write_le(&mut out[16..32]);
-        (x.c + m.unpack(self.c)).write_le(&mut out[32..48]);
-        (x.d + m.unpack(self.d)).write_le(&mut out[48..64]);
-    }
-
-    /// Add one to the block counter (no overflow check).
-    #[inline(always)]
-    fn inc_block_ct<M: Machine>(&mut self, m: M) {
-        let d0: M::u64x2 = m.unpack(self.d);
-        self.d = (d0 + m.vec([1, 0])).into();
-    }
-}
-
-mod mach {
-    use super::*;
-    #[inline(always)]
-    pub fn refill_wide_impl<Mach: Machine>(
-        m: Mach,
-        state: &mut ChaCha,
-        drounds: u32,
-        out: &mut [u8; BUFSZ],
-    ) {
-        let k = m.vec([0x6170_7865, 0x3320_646e, 0x7962_2d32, 0x6b20_6574]);
-        // TODO: endian
-        let inc = m.vec([1, 0]);
-        let d0: Mach::u64x2 = m.unpack(state.d);
-        let d1 = d0 + inc;
-        let d2 = d1 + inc;
-        let d3 = d2 + inc;
-        let b = m.unpack(state.b);
-        let c = m.unpack(state.c);
-        let mut x = State {
-            a: Mach::u32x4x4::from_lanes([k, k, k, k]),
-            b: Mach::u32x4x4::from_lanes([b, b, b, b]),
-            c: Mach::u32x4x4::from_lanes([c, c, c, c]),
-            d: m.unpack(Mach::u64x2x4::from_lanes([d0, d1, d2, d3]).into()),
-        };
-        for _ in 0..drounds {
-            x = round(x);
-            x = undiagonalize(round(diagonalize(x)));
-        }
-        let inc = m.vec([1, 0]);
-        let d0: Mach::u64x2 = m.unpack(state.d);
-        let d1 = d0 + inc;
-        let d2 = d1 + inc;
-        let d3 = d2 + inc;
-        let d4 = d3 + inc;
-        let d1: Mach::u32x4 = m.unpack(d1.into());
-        let d2: Mach::u32x4 = m.unpack(d2.into());
-        let d3: Mach::u32x4 = m.unpack(d3.into());
-        let (a, b, c, d) = (
-            x.a.to_lanes(),
-            x.b.to_lanes(),
-            x.c.to_lanes(),
-            x.d.to_lanes(),
-        );
-        let sb = m.unpack(state.b);
-        let sc = m.unpack(state.c);
-        let sd = [m.unpack(state.d), d1, d2, d3];
-        state.d = d4.into();
-        let mut words = out.chunks_exact_mut(16);
-        for ((((&a, &b), &c), &d), &sd) in a.iter().zip(&b).zip(&c).zip(&d).zip(&sd) {
-            (a + k).write_le(words.next().unwrap());
-            (b + sb).write_le(words.next().unwrap());
-            (c + sc).write_le(words.next().unwrap());
-            (d + sd).write_le(words.next().unwrap());
-        }
-    }
-}
-/// XXX unstable, only exposed for benchmarks
-pub use self::mach::refill_wide_impl;
-
-dispatch!(m, Mach, {
-    fn refill_wide(state: &mut ChaCha, drounds: u32, out: &mut [u8; BUFSZ]) {
-        refill_wide_impl(m, state, drounds, out);
-    }
-});
-
-/// Refill the buffer from a single-block round, updating the block count.
-dispatch_light128!(m, Mach, {
-    fn refill_narrow(state: &mut ChaCha, drounds: u32, out: &mut [u8; BLOCK]) {
-        let x = refill_narrow_rounds(state, drounds);
-        let x = State {
-            a: m.unpack(x.a),
-            b: m.unpack(x.b),
-            c: m.unpack(x.c),
-            d: m.unpack(x.d),
-        };
-        state.output_narrow(m, x, out);
-        state.inc_block_ct(m);
-    }
-});
-
-/// Single-block, rounds-only; shared by try_apply_keystream for tails shorter than BUFSZ
-/// and XChaCha's setup step.
-dispatch!(m, Mach, {
-    fn refill_narrow_rounds(state: &mut ChaCha, drounds: u32) -> State<vec128_storage> {
-        let k: Mach::u32x4 = m.vec([0x6170_7865, 0x3320_646e, 0x7962_2d32, 0x6b20_6574]);
-        let mut x = State {
-            a: k,
-            b: m.unpack(state.b),
-            c: m.unpack(state.c),
-            d: m.unpack(state.d),
-        };
-        for _ in 0..drounds {
-            x = round(x);
-            x = undiagonalize(round(diagonalize(x)));
-        }
-        State {
-            a: x.a.into(),
-            b: x.b.into(),
-            c: x.c.into(),
-            d: x.d.into(),
-        }
-    }
-});
-
-mod chacha_any {
-    use super::*;
-    #[derive(Clone)]
-    pub struct ChaCha {
-        pub b: vec128_storage,
-        pub c: vec128_storage,
-        pub d: vec128_storage,
-    }
-
-    #[derive(Clone)]
-    pub struct Buffer {
-        pub state: ChaCha,
-        pub out: [u8; BLOCK],
-        pub have: i8,
-        pub len: u64,
-        pub fresh: bool,
-    }
-
-    #[derive(Default)]
-    pub struct X;
-    #[derive(Default)]
-    pub struct O;
-    #[derive(Clone)]
-    pub struct ChaChaAny<NonceSize, Rounds, IsX> {
-        pub state: Buffer,
-        pub _nonce_size: NonceSize,
-        pub _rounds: Rounds,
-        pub _is_x: IsX,
-    }
-}
-/// XXX unstable, only exposed for benchmarks
-pub use self::chacha_any::ChaCha as ChaChaState;
-use self::chacha_any::*;
-
-trait AsBool {
-    const BOOL: bool;
-}
-struct WideEnabled;
-impl AsBool for WideEnabled {
-    const BOOL: bool = true;
-}
-#[cfg(test)]
-struct WideDisabled;
-#[cfg(test)]
-impl AsBool for WideDisabled {
-    const BOOL: bool = false;
-}
-
-impl Buffer {
-    fn try_apply_keystream<EnableWide: AsBool>(
-        &mut self,
-        mut data: &mut [u8],
-        drounds: u32,
-    ) -> Result<(), ()> {
-        // Lazy fill: after a seek() we may be partway into a block we don't have yet.
-        // We can do this before the overflow check because this is not an effect of the current
-        // operation.
-        if self.have < 0 {
-            refill_narrow(&mut self.state, drounds, &mut self.out);
-            self.have += BLOCK as i8;
-            // checked in seek()
-            self.len -= 1;
-        }
-        let mut have = self.have as usize;
-        let have_ready = cmp::min(have, data.len());
-        // Check if the requested position would wrap the block counter. Use self.fresh as an extra
-        // bit to distinguish the initial state from the valid state with no blocks left.
-        let datalen = (data.len() - have_ready) as u64;
-        let blocks_needed = datalen / BLOCK64 + u64::from(datalen % BLOCK64 != 0);
-        let (l, o) = self.len.overflowing_sub(blocks_needed);
-        if o && !self.fresh {
-            return Err(());
-        }
-        self.len = l;
-        self.fresh &= blocks_needed == 0;
-        // If we have data in the buffer, use that first.
-        let (d0, d1) = data.split_at_mut(have_ready);
-        for (data_b, key_b) in d0.iter_mut().zip(&self.out[(BLOCK - have)..]) {
-            *data_b ^= *key_b;
-        }
-        data = d1;
-        have -= have_ready;
-        // Process wide chunks.
-        if EnableWide::BOOL {
-            let (d0, d1) = data.split_at_mut(data.len() & !(BUFSZ - 1));
-            for dd in d0.chunks_exact_mut(BUFSZ) {
-                let mut buf = [0; BUFSZ];
-                refill_wide(&mut self.state, drounds, &mut buf);
-                for (data_b, key_b) in dd.iter_mut().zip(buf.iter()) {
-                    *data_b ^= *key_b;
-                }
-            }
-            data = d1;
-        }
-        // Handle the tail a block at a time so we'll have storage for any leftovers.
-        for dd in data.chunks_mut(BLOCK) {
-            refill_narrow(&mut self.state, drounds, &mut self.out);
-            for (data_b, key_b) in dd.iter_mut().zip(self.out.iter()) {
-                *data_b ^= *key_b;
-            }
-            have = BLOCK - dd.len();
-        }
-        self.have = have as i8;
-        Ok(())
-    }
-}
-
-dispatch_light128!(m, Mach, {
-    fn seek64(buf: &mut Buffer, ct: u64) {
-        let blockct = ct / BLOCK64;
-        buf.len = BIG_LEN.wrapping_sub(blockct);
-        buf.fresh = blockct == 0;
-        buf.have = -((ct % BLOCK64) as i8);
-        buf.state.seek64(m, blockct);
-    }
-});
-
-dispatch_light128!(m, Mach, {
-    fn set_stream_param(state: &mut ChaCha, param: u32, value: u64) {
-        let d: Mach::u64x2 = m.unpack(state.d);
-        state.d = d.insert(value, param).into();
-    }
-});
-
-dispatch_light128!(m, Mach, {
-    fn get_stream_param(state: &ChaCha, param: u32) -> u64 {
-        let d: Mach::u64x2 = m.unpack(state.d);
-        d.extract(param).into()
-    }
-});
-
-dispatch_light128!(m, Mach, {
-    fn seek32(buf: &mut Buffer, ct: u64) {
-        let blockct = ct / BLOCK64;
-        assert!(blockct < SMALL_LEN || (blockct == SMALL_LEN && ct % BLOCK64 == 0));
-        buf.len = SMALL_LEN - blockct;
-        buf.have = -((ct % BLOCK64) as i8);
-        buf.state.seek32(m, blockct as u32);
-    }
-});
-
-#[cfg(test)]
-impl<NonceSize, Rounds: Unsigned, IsX> ChaChaAny<NonceSize, Rounds, IsX> {
-    pub fn try_apply_keystream_narrow(&mut self, data: &mut [u8]) -> Result<(), ()> {
-        self.state
-            .try_apply_keystream::<WideDisabled>(data, Rounds::U32)
-    }
-}
-
-dispatch_light128!(m, Mach, {
-    fn init_chacha(key: &GenericArray<u8, U32>, nonce: &[u8]) -> ChaCha {
-        let ctr_nonce = [
-            0,
-            if nonce.len() == 12 {
-                LE::read_u32(&nonce[0..4])
-            } else {
-                0
-            },
-            LE::read_u32(&nonce[nonce.len() - 8..nonce.len() - 4]),
-            LE::read_u32(&nonce[nonce.len() - 4..]),
-        ];
-        let key0: Mach::u32x4 = m.read_le(&key[..16]);
-        let key1: Mach::u32x4 = m.read_le(&key[16..]);
-        ChaCha {
-            b: key0.into(),
-            c: key1.into(),
-            d: ctr_nonce.into(),
-        }
-    }
-});
-
-dispatch_light128!(m, Mach, {
-    fn init_chacha_x(
-        key: &GenericArray<u8, U32>,
-        nonce: &GenericArray<u8, U24>,
-        rounds: u32,
-    ) -> ChaCha {
-        let key0: Mach::u32x4 = m.read_le(&key[..16]);
-        let key1: Mach::u32x4 = m.read_le(&key[16..]);
-        let nonce0: Mach::u32x4 = m.read_le(&nonce[..16]);
-        let mut state = ChaCha {
-            b: key0.into(),
-            c: key1.into(),
-            d: nonce0.into(),
-        };
-        let x = refill_narrow_rounds(&mut state, rounds);
-        let ctr_nonce1 = [
-            0,
-            0,
-            LE::read_u32(&nonce[16..20]),
-            LE::read_u32(&nonce[20..24]),
-        ];
-        state.b = x.a;
-        state.c = x.d;
-        state.d = ctr_nonce1.into();
-        state
-    }
-});
-
-impl<NonceSize, Rounds> ChaChaAny<NonceSize, Rounds, O>
-where
-    NonceSize: Unsigned + ArrayLength<u8> + Default,
-    Rounds: Default,
-{
-    #[inline]
-    fn new(key: &GenericArray<u8, U32>, nonce: &GenericArray<u8, NonceSize>) -> Self {
-        let nonce_len = nonce.len();
-        ChaChaAny {
-            state: Buffer {
-                state: init_chacha(key, nonce),
-                out: [0; BLOCK],
-                have: 0,
-                len: if nonce_len == 12 { SMALL_LEN } else { BIG_LEN },
-                fresh: nonce_len != 12,
-            },
-            _nonce_size: Default::default(),
-            _rounds: Default::default(),
-            _is_x: Default::default(),
-        }
-    }
-}
-
-impl<Rounds: Unsigned + Default> ChaChaAny<U24, Rounds, X> {
-    fn new(key: &GenericArray<u8, U32>, nonce: &GenericArray<u8, U24>) -> Self {
-        ChaChaAny {
-            state: Buffer {
-                state: init_chacha_x(key, nonce, Rounds::U32),
-                out: [0; BLOCK],
-                have: 0,
-                len: BIG_LEN,
-                fresh: true,
-            },
-            _nonce_size: Default::default(),
-            _rounds: Default::default(),
-            _is_x: Default::default(),
-        }
-    }
-}
-
-impl<NonceSize: Unsigned, Rounds, IsX> ChaChaAny<NonceSize, Rounds, IsX> {
-    #[inline(always)]
-    fn seek(&mut self, ct: u64) {
-        if NonceSize::U32 != 12 {
-            seek64(&mut self.state, ct);
-        } else {
-            seek32(&mut self.state, ct);
-        }
-    }
-}
-
-impl<NonceSize, Rounds: Unsigned, IsX> ChaChaAny<NonceSize, Rounds, IsX> {
-    #[inline]
-    fn try_apply_keystream(&mut self, data: &mut [u8]) -> Result<(), ()> {
-        self.state
-            .try_apply_keystream::<WideEnabled>(data, Rounds::U32)
-    }
-    #[inline]
-    fn apply_keystream(&mut self, data: &mut [u8]) {
-        self.try_apply_keystream(data).expect("keystream exhausted")
-    }
-}
-
-#[cfg(feature = "rustcrypto_api")]
-mod rustcrypto_impl {
-    use super::*;
-    use stream_cipher::{LoopError, NewStreamCipher, SyncStreamCipher, SyncStreamCipherSeek};
-
-    impl<NonceSize, Rounds> NewStreamCipher for ChaChaAny<NonceSize, Rounds, O>
-    where
-        NonceSize: Unsigned + ArrayLength<u8> + Default,
-        Rounds: Default,
-    {
-        type KeySize = U32;
-        type NonceSize = NonceSize;
-        #[inline]
-        fn new(
-            key: &GenericArray<u8, Self::KeySize>,
-            nonce: &GenericArray<u8, Self::NonceSize>,
-        ) -> Self {
-            Self::new(key, nonce)
-        }
-    }
-
-    impl<Rounds: Unsigned + Default> NewStreamCipher for ChaChaAny<U24, Rounds, X> {
-        type KeySize = U32;
-        type NonceSize = U24;
-        #[inline]
-        fn new(
-            key: &GenericArray<u8, Self::KeySize>,
-            nonce: &GenericArray<u8, Self::NonceSize>,
-        ) -> Self {
-            Self::new(key, nonce)
-        }
-    }
-
-    impl<NonceSize: Unsigned, Rounds, IsX> SyncStreamCipherSeek for ChaChaAny<NonceSize, Rounds, IsX> {
-        #[inline]
-        fn current_pos(&self) -> u64 {
-            unimplemented!()
-        }
-        #[inline(always)]
-        fn seek(&mut self, ct: u64) {
-            Self::seek(self, ct)
-        }
-    }
-
-    impl<NonceSize, Rounds: Unsigned, IsX> SyncStreamCipher for ChaChaAny<NonceSize, Rounds, IsX> {
-        #[inline]
-        fn try_apply_keystream(&mut self, data: &mut [u8]) -> Result<(), LoopError> {
-            Self::try_apply_keystream(self, data).map_err(|_| LoopError)
-        }
-    }
-}
-
-/// IETF RFC 7539 ChaCha. Unsuitable for messages longer than 256 GiB.
-pub type Ietf = ChaChaAny<U12, U10, O>;
-/// ChaCha20, as used in several standards; from Bernstein's original publication.
-pub type ChaCha20 = ChaChaAny<U8, U10, O>;
-/// Similar to ChaCha20, but with fewer rounds for higher performance.
-pub type ChaCha12 = ChaChaAny<U8, U6, O>;
-/// Similar to ChaCha20, but with fewer rounds for higher performance.
-pub type ChaCha8 = ChaChaAny<U8, U4, O>;
-/// Constructed analogously to XSalsa20; mixes during initialization to support both a long nonce
-/// and a full-length (64-bit) block counter.
-pub type XChaCha20 = ChaChaAny<U24, U10, X>;
+mod crypto;
 
 #[cfg(feature = "provide_rand_backend")]
-mod rand_impl {
-    use super::generic_array::typenum::{U10, U64};
-    use super::*;
-    use core::marker::PhantomData;
-    use rand_core::block::{BlockRng, BlockRngCore};
-    use rand_core::{CryptoRng, Error, RngCore, SeedableRng};
-
-    /// The core of `ChaChaRng`, used with `BlockRng`.
-    #[derive(Clone)]
-    pub struct ChaChaCore<Rounds> {
-        state: ChaCha,
-        _rounds: PhantomData<Rounds>,
-    }
-
-    // Custom Debug implementation that does not expose the internal state
-    impl fmt::Debug for ChaChaCore {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "ChaChaCore {{}}")
-        }
-    }
-
-    impl<Rounds: Unsigned> BlockRngCore for ChaChaCore<Rounds> {
-        type Item = u32;
-        type Results = GenericArray<Self::Item, U64>;
-        #[inline]
-        fn generate(&mut self, r: &mut Self::Results) {
-            // Fill slice of words by writing to equivalent slice of bytes, then fixing endianness.
-            refill_wide(&mut self.state, Rounds::U32, unsafe {
-                core::mem::transmute::<&mut GenericArray<u32, U64>, &mut [u8; 256]>(&mut *r)
-            });
-            for x in r {
-                *x = x.to_le();
-            }
-        }
-    }
-
-    /// A cryptographically secure random number generator that uses the ChaCha algorithm.
-    ///
-    /// ChaCha is a stream cipher designed by Daniel J. Bernstein[^1], that we use as an RNG. It is
-    /// an improved variant of the Salsa20 cipher family, which was selected as one of the "stream
-    /// ciphers suitable for widespread adoption" by eSTREAM[^2].
-    ///
-    /// ChaCha uses add-rotate-xor (ARX) operations as its basis. These are safe against timing
-    /// attacks, although that is mostly a concern for ciphers and not for RNGs. We provide a SIMD
-    /// implementation to support high throughput on a variety of common hardware platforms.
-    ///
-    /// With the ChaCha algorithm it is possible to choose the number of rounds the core algorithm
-    /// should run. The number of rounds is a tradeoff between performance and security, where 8
-    /// rounds is the minimum potentially secure configuration, and 20 rounds is widely used as a
-    /// conservative choice. We support type-level configuration of the round count with `typenum`;
-    /// note that the implementation operates in double-rounds, so the parameter must be specified
-    /// as half the desired round count.
-    ///
-    /// We use a 64-bit counter and 64-bit stream identifier as in Bernstein's implementation[^1]
-    /// except that we use a stream identifier in place of a nonce. A 64-bit counter over 64-byte
-    /// (16 word) blocks allows 1 ZiB of output before cycling, and the stream identifier allows
-    /// 2<sup>64</sup> unique streams of output per seed. Both counter and stream are initialized
-    /// to zero but may be set via [`set_word_pos`] and [`set_stream`].
-    ///
-    /// The word layout is:
-    ///
-    /// ```text
-    /// constant  constant  constant  constant
-    /// seed      seed      seed      seed
-    /// seed      seed      seed      seed
-    /// counter   counter   stream_id stream_id
-    /// ```
-    ///
-    /// This implementation uses an output buffer of sixteen `u32` words, and uses
-    /// [`BlockRng`] to implement the [`RngCore`] methods.
-    ///
-    /// [^1]: D. J. Bernstein, [*ChaCha, a variant of Salsa20*](
-    ///       https://cr.yp.to/chacha.html)
-    ///
-    /// [^2]: [eSTREAM: the ECRYPT Stream Cipher Project](
-    ///       http://www.ecrypt.eu.org/stream/)
-    ///
-    /// [`set_word_pos`]: ChaChaRng::set_word_pos
-    /// [`set_stream`]: ChaChaRng::set_stream
-    #[derive(Clone, Debug)]
-    pub struct ChaChaRng<Rounds: Unsigned> {
-        rng: BlockRng<ChaChaCore<Rounds>>,
-    }
-
-    impl<Rounds: Unsigned> SeedableRng for ChaChaRng<Rounds> {
-        type Seed = [u8; 32];
-        #[inline]
-        fn from_seed(seed: Self::Seed) -> Self {
-            let core = ChaChaCore {
-                state: init_chacha(GenericArray::from_slice(&seed), &[0u8; 8]),
-                _rounds: PhantomData,
-            };
-            Self {
-                rng: BlockRng::new(core),
-            }
-        }
-    }
-
-    impl<Rounds: Unsigned> RngCore for ChaChaRng<Rounds> {
-        #[inline]
-        fn next_u32(&mut self) -> u32 {
-            self.rng.next_u32()
-        }
-        #[inline]
-        fn next_u64(&mut self) -> u64 {
-            self.rng.next_u64()
-        }
-        #[inline]
-        fn fill_bytes(&mut self, bytes: &mut [u8]) {
-            self.rng.fill_bytes(bytes)
-        }
-        #[inline]
-        fn try_fill_bytes(&mut self, bytes: &mut [u8]) -> Result<(), Error> {
-            self.rng.try_fill_bytes(bytes)
-        }
-    }
-
-    const STREAM_PARAM_NONCE: u32 = 1;
-    const STREAM_PARAM_BLOCK: u32 = 0;
-    impl<Rounds: Unsigned> ChaChaRng<Rounds> {
-        // The buffer is a 4-block window, i.e. it is always at a block-aligned position in the
-        // stream but if the stream has been seeked it may not be self-aligned.
-
-        /// Get the offset from the start of the stream, in 32-bit words.
-        ///
-        /// Since the generated blocks are 16 words (2<sup>4</sup>) long and the
-        /// counter is 64-bits, the offset is a 68-bit number. Sub-word offsets are
-        /// not supported, hence the result can simply be multiplied by 4 to get a
-        /// byte-offset.
-        #[inline]
-        pub fn get_word_pos(&self) -> u128 {
-            let block = u128::from(get_stream_param(&self.rng.core.state, STREAM_PARAM_BLOCK));
-            (block << 4) + self.rng.index() as u128
-        }
-
-        /// Set the offset from the start of the stream, in 32-bit words.
-        ///
-        /// As with `get_word_pos`, we use a 68-bit number. Since the generator
-        /// simply cycles at the end of its period (1 ZiB), we ignore the upper
-        /// 60 bits.
-        #[inline]
-        pub fn set_word_pos(&mut self, word_offset: u128) {
-            let block = (word_offset >> 4) as u64;
-            set_stream_param(&mut self.rng.core.state, STREAM_PARAM_BLOCK, block);
-            self.rng.generate_and_set((word_offset & 15) as usize);
-        }
-
-        /// Set the stream number.
-        ///
-        /// This is initialized to zero; 2<sup>64</sup> unique streams of output
-        /// are available per seed/key.
-        ///
-        /// Note that in order to reproduce ChaCha output with a specific 64-bit
-        /// nonce, one can convert that nonce to a `u64` in little-endian fashion
-        /// and pass to this function. In theory a 96-bit nonce can be used by
-        /// passing the last 64-bits to this function and using the first 32-bits as
-        /// the most significant half of the 64-bit counter (which may be set
-        /// indirectly via `set_word_pos`), but this is not directly supported.
-        #[inline]
-        pub fn set_stream(&mut self, stream: u64) {
-            set_stream_param(&mut self.rng.core.state, STREAM_PARAM_NONCE, stream);
-            if self.rng.index() != 64 {
-                let wp = self.get_word_pos();
-                self.set_word_pos(wp);
-            }
-        }
-    }
-
-    impl<Rounds: Unsigned> CryptoRng for ChaChaRng<Rounds> {}
-
-    pub type ChaCha20Rng = ChaChaRng<U10>;
-
-    impl From<ChaChaCore> for ChaChaRng {
-        fn from(core: ChaChaCore) -> Self {
-            ChaChaRng {
-                rng: BlockRng::new(core),
-            }
-        }
-    }
-
-    #[cfg(test)]
-    mod test {
-        use super::ChaCha20Rng as ChaChaRng;
-        use rand_core::{RngCore, SeedableRng};
-
-        #[test]
-        fn test_chacha_construction() {
-            let seed = [
-                0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0,
-                0, 0, 0, 0,
-            ];
-            let mut rng1 = ChaChaRng::from_seed(seed);
-            assert_eq!(rng1.next_u32(), 137206642);
-
-            let mut rng2 = ChaChaRng::from_rng(rng1).unwrap();
-            assert_eq!(rng2.next_u32(), 1325750369);
-        }
-
-        #[test]
-        fn test_chacha_true_values_a() {
-            // Test vectors 1 and 2 from
-            // https://tools.ietf.org/html/draft-nir-cfrg-chacha20-poly1305-04
-            let seed = [0u8; 32];
-            let mut rng = ChaChaRng::from_seed(seed);
-
-            let mut results = [0u32; 16];
-            for i in results.iter_mut() {
-                *i = rng.next_u32();
-            }
-            let expected = [
-                0xade0b876, 0x903df1a0, 0xe56a5d40, 0x28bd8653, 0xb819d2bd, 0x1aed8da0, 0xccef36a8,
-                0xc70d778b, 0x7c5941da, 0x8d485751, 0x3fe02477, 0x374ad8b8, 0xf4b8436a, 0x1ca11815,
-                0x69b687c3, 0x8665eeb2,
-            ];
-            assert_eq!(results, expected);
-
-            for i in results.iter_mut() {
-                *i = rng.next_u32();
-            }
-            let expected = [
-                0xbee7079f, 0x7a385155, 0x7c97ba98, 0x0d082d73, 0xa0290fcb, 0x6965e348, 0x3e53c612,
-                0xed7aee32, 0x7621b729, 0x434ee69c, 0xb03371d5, 0xd539d874, 0x281fed31, 0x45fb0a51,
-                0x1f0ae1ac, 0x6f4d794b,
-            ];
-            assert_eq!(results, expected);
-        }
-
-        #[test]
-        fn test_chacha_true_values_b() {
-            // Test vector 3 from
-            // https://tools.ietf.org/html/draft-nir-cfrg-chacha20-poly1305-04
-            let seed = [
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 1,
-            ];
-            let mut rng = ChaChaRng::from_seed(seed);
-
-            // Skip block 0
-            for _ in 0..16 {
-                rng.next_u32();
-            }
-
-            let mut results = [0u32; 16];
-            for i in results.iter_mut() {
-                *i = rng.next_u32();
-            }
-            let expected = [
-                0x2452eb3a, 0x9249f8ec, 0x8d829d9b, 0xddd4ceb1, 0xe8252083, 0x60818b01, 0xf38422b8,
-                0x5aaa49c9, 0xbb00ca8e, 0xda3ba7b4, 0xc4b592d1, 0xfdf2732f, 0x4436274e, 0x2561b3c8,
-                0xebdd4aa6, 0xa0136c00,
-            ];
-            assert_eq!(results, expected);
-        }
-
-        #[test]
-        #[cfg(all(rustc_1_26, not(target_os = "emscripten")))]
-        fn test_chacha_true_values_c() {
-            // Test vector 4 from
-            // https://tools.ietf.org/html/draft-nir-cfrg-chacha20-poly1305-04
-            let seed = [
-                0, 0xff, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 0, 0,
-            ];
-            let expected = [
-                0xfb4dd572, 0x4bc42ef1, 0xdf922636, 0x327f1394, 0xa78dea8f, 0x5e269039, 0xa1bebbc1,
-                0xcaf09aae, 0xa25ab213, 0x48a6b46c, 0x1b9d9bcb, 0x092c5be6, 0x546ca624, 0x1bec45d5,
-                0x87f47473, 0x96f0992e,
-            ];
-            let expected_end = 3 * 16;
-            let mut results = [0u32; 16];
-
-            // Test block 2 by skipping block 0 and 1
-            let mut rng1 = ChaChaRng::from_seed(seed);
-            for _ in 0..32 {
-                rng1.next_u32();
-            }
-            for i in results.iter_mut() {
-                *i = rng1.next_u32();
-            }
-            assert_eq!(results, expected);
-            assert_eq!(rng1.get_word_pos(), expected_end);
-
-            // Test block 2 by using `set_word_pos`
-            let mut rng2 = ChaChaRng::from_seed(seed);
-            rng2.set_word_pos(2 * 16);
-            for i in results.iter_mut() {
-                *i = rng2.next_u32();
-            }
-            assert_eq!(results, expected);
-            assert_eq!(rng2.get_word_pos(), expected_end);
-
-            // Test skipping behaviour with other types
-            let mut buf = [0u8; 32];
-            rng2.fill_bytes(&mut buf[..]);
-            assert_eq!(rng2.get_word_pos(), expected_end + 8);
-            rng2.fill_bytes(&mut buf[0..25]);
-            assert_eq!(rng2.get_word_pos(), expected_end + 15);
-            rng2.next_u64();
-            assert_eq!(rng2.get_word_pos(), expected_end + 17);
-            rng2.next_u32();
-            rng2.next_u64();
-            assert_eq!(rng2.get_word_pos(), expected_end + 20);
-            rng2.fill_bytes(&mut buf[0..1]);
-            assert_eq!(rng2.get_word_pos(), expected_end + 21);
-        }
-
-        #[test]
-        fn test_chacha_multiple_blocks() {
-            let seed = [
-                0, 0, 0, 0, 1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0, 0, 4, 0, 0, 0, 5, 0, 0, 0, 6, 0, 0, 0,
-                7, 0, 0, 0,
-            ];
-            let mut rng = ChaChaRng::from_seed(seed);
-
-            // Store the 17*i-th 32-bit word,
-            // i.e., the i-th word of the i-th 16-word block
-            let mut results = [0u32; 16];
-            for i in results.iter_mut() {
-                *i = rng.next_u32();
-                for _ in 0..16 {
-                    rng.next_u32();
-                }
-            }
-            let expected = [
-                0xf225c81a, 0x6ab1be57, 0x04d42951, 0x70858036, 0x49884684, 0x64efec72, 0x4be2d186,
-                0x3615b384, 0x11cfa18e, 0xd3c50049, 0x75c775f6, 0x434c6530, 0x2c5bad8f, 0x898881dc,
-                0x5f1c86d9, 0xc1f8e7f4,
-            ];
-            assert_eq!(results, expected);
-        }
-
-        #[test]
-        fn test_chacha_true_bytes() {
-            let seed = [0u8; 32];
-            let mut rng = ChaChaRng::from_seed(seed);
-            let mut results = [0u8; 32];
-            rng.fill_bytes(&mut results);
-            let expected = [
-                118, 184, 224, 173, 160, 241, 61, 144, 64, 93, 106, 229, 83, 134, 189, 40, 189,
-                210, 25, 184, 160, 141, 237, 26, 168, 54, 239, 204, 139, 119, 13, 199,
-            ];
-            assert_eq!(results, expected);
-        }
-
-        #[test]
-        fn test_chacha_nonce() {
-            // Test vector 5 from
-            // https://tools.ietf.org/html/draft-nir-cfrg-chacha20-poly1305-04
-            // Although we do not support setting a nonce, we try it here anyway so
-            // we can use this test vector.
-            let seed = [0u8; 32];
-            let mut rng = ChaChaRng::from_seed(seed);
-            // 96-bit nonce in LE order is: 0,0,0,0, 0,0,0,0, 0,0,0,2
-            rng.set_stream(2u64 << (24 + 32));
-
-            let mut results = [0u32; 16];
-            for i in results.iter_mut() {
-                *i = rng.next_u32();
-            }
-            let expected = [
-                0x374dc6c2, 0x3736d58c, 0xb904e24a, 0xcd3f93ef, 0x88228b1a, 0x96a4dfb3, 0x5b76ab72,
-                0xc727ee54, 0x0e0e978a, 0xf3145c95, 0x1b748ea8, 0xf786c297, 0x99c28f5f, 0x628314e8,
-                0x398a19fa, 0x6ded1b53,
-            ];
-            assert_eq!(results, expected);
-        }
-
-        #[test]
-        fn test_chacha_clone_streams() {
-            let seed = [
-                0, 0, 0, 0, 1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0, 0, 4, 0, 0, 0, 5, 0, 0, 0, 6, 0, 0, 0,
-                7, 0, 0, 0,
-            ];
-            let mut rng = ChaChaRng::from_seed(seed);
-            let mut clone = rng.clone();
-            for _ in 0..16 {
-                assert_eq!(rng.next_u64(), clone.next_u64());
-            }
-
-            rng.set_stream(51);
-            for _ in 0..7 {
-                assert!(rng.next_u32() != clone.next_u32());
-            }
-            clone.set_stream(51); // switch part way through block
-            for _ in 7..16 {
-                assert_eq!(rng.next_u32(), clone.next_u32());
-            }
-        }
-    }
-}
+mod rand_impl;
 #[cfg(feature = "provide_rand_backend")]
-pub use rand_impl::{ChaCha20Rng, ChaChaCore, ChaChaRng};
+pub use self::rand_impl::{ChaCha20Rng, ChaChaCore, ChaChaRng};
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn chacha20_case_1() {
-        let key = hex!("fa44478c59ca70538e3549096ce8b523232c50d9e8e8d10c203ef6c8d07098a5");
-        let nonce = hex!("8d3a0d6d7827c007");
-        let expected = hex!("
-                1546a547ff77c5c964e44fd039e913c6395c8f19d43efaa880750f6687b4e6e2d8f42f63546da2d133b5aa2f1ef3f218b6c72943089e4012
-                210c2cbed0e8e93498a6825fc8ff7a504f26db33b6cbe36299436244c9b2eff88302c55933911b7d5dea75f2b6d4761ba44bb6f814c9879d
-                2ba2ac8b178fa1104a368694872339738ffb960e33db39efb8eaef885b910eea078e7a1feb3f8185dafd1455b704d76da3a0ce4760741841
-                217bba1e4ece760eaf68617133431feb806c061173af6b8b2a23be90c5d145cc258e3c119aab2800f0c7bc1959dae75481712cab731b7dfd
-                783fa3a228f9968aaea68f36a92f43c9b523337a55b97bcaf5f5774447bf41e8");
-        let mut state = ChaCha20::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let offset = 0x3fffffff70u64;
-        assert!((offset >> 38) != ((offset + 240) >> 38)); // This will overflow the small word of the counter
-        state.seek(offset);
-        let mut result = [0; 256];
-        state.apply_keystream(&mut result);
-        assert_eq!(&expected[..], &result[..]);
-    }
-
-    #[test]
-    fn chacha12_case_1() {
-        let key = hex!("27fc120b013b829f1faeefd1ab417e8662f43e0d73f98de866e346353180fdb7");
-        let nonce = hex!("db4b4a41d8df18aa");
-        let expected = hex!("
-                5f3c8c190a78ab7fe808cae9cbcb0a9837c893492d963a1c2eda6c1558b02c83fc02a44cbbb7e6204d51d1c2430e9c0b58f2937bf593840c
-                850bda9051a1f051ddf09d2a03ebf09f01bdba9da0b6da791b2e645641047d11ebf85087d4de5c015fddd044");
-        let mut state = ChaCha12::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let mut result = [0u8; 100];
-        state.apply_keystream(&mut result);
-        assert_eq!(&expected[..], &result[..]);
-    }
-
-    #[test]
-    fn chacha8_case_1() {
-        let key = hex!("641aeaeb08036b617a42cf14e8c5d2d115f8d7cb6ea5e28b9bfaf83e038426a7");
-        let nonce = hex!("a14a1168271d459b");
-        let mut state = ChaCha8::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let expected = hex!(
-        "1721c044a8a6453522dddb3143d0be3512633ca3c79bf8ccc3594cb2c2f310f7bd544f55ce0db38123412d6c45207d5cf9af0c6c680cce1f
-        7e43388d1b0346b7133c59fd6af4a5a568aa334ccdc38af5ace201df84d0a3ca225494ca6209345fcf30132e");
-        let mut result = [0u8; 100];
-        state.apply_keystream(&mut result);
-        assert_eq!(&expected[..], &result[..]);
-    }
-
-    #[test]
-    fn test_ietf() {
-        let key = hex!("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f");
-        let nonce = hex!("000000090000004a00000000");
-        let expected = hex!(
-            "
-            10f1e7e4d13b5915500fdd1fa32071c4c7d1f4c733c068030422aa9ac3d46c4e
-            d2826446079faa0914c2d705d98b02a2b5129cd1de164eb9cbd083e8a2503c4e"
-        );
-        let mut state = Ietf::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let mut result = [0; 64];
-        state.seek(64);
-        state.apply_keystream(&mut result);
-        assert_eq!(&expected[..], &result[..]);
-    }
-
-    #[test]
-    fn rfc_7539_case_1() {
-        let key = hex!("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f");
-        let nonce = hex!("000000090000004a00000000");
-        let mut state = Ietf::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let mut result = [0; 128];
-        state.apply_keystream(&mut result);
-        let expected = hex!(
-            "10f1e7e4d13b5915500fdd1fa32071c4c7d1f4c733c068030422aa9ac3d46c4e
-            d2826446079faa0914c2d705d98b02a2b5129cd1de164eb9cbd083e8a2503c4e"
-        );
-        assert_eq!(&expected[..], &result[64..]);
-    }
-
-    #[test]
-    fn rfc_7539_case_2() {
-        let key = hex!("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f");
-        let nonce = hex!("000000000000004a00000000");
-        let mut state = Ietf::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let plaintext = b"Ladies and Gentlemen of the class of '99: If I could offer you only one tip for the future, sunscreen would be it.";
-        let mut buf = [0u8; 178];
-        buf[64..].copy_from_slice(plaintext);
-        state.apply_keystream(&mut buf);
-        let expected = hex!("
-            6e2e359a2568f98041ba0728dd0d6981e97e7aec1d4360c20a27afccfd9fae0bf91b65c5524733ab8f593dabcd62b3571639d624e65152ab
-            8f530c359f0861d807ca0dbf500d6a6156a38e088a22b65e52bc514d16ccf806818ce91ab77937365af90bbf74a35be6b40b8eedf2785e42
-            874d");
-        assert_eq!(&expected[..], &buf[64..]);
-    }
-
-    #[test]
-    fn rfc_7539_case_2_chunked() {
-        let key = hex!("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f");
-        let nonce = hex!("000000000000004a00000000");
-        let mut state = Ietf::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let plaintext = b"Ladies and Gentlemen of the class of '99: If I could offer you only one tip for the future, sunscreen would be it.";
-        let mut buf = [0u8; 178];
-        buf[64..].copy_from_slice(plaintext);
-        state.apply_keystream(&mut buf[..40]);
-        state.apply_keystream(&mut buf[40..78]);
-        state.apply_keystream(&mut buf[78..79]);
-        state.apply_keystream(&mut buf[79..128]);
-        state.apply_keystream(&mut buf[128..]);
-        let expected = hex!("
-            6e2e359a2568f98041ba0728dd0d6981e97e7aec1d4360c20a27afccfd9fae0bf91b65c5524733ab8f593dabcd62b3571639d624e65152ab
-            8f530c359f0861d807ca0dbf500d6a6156a38e088a22b65e52bc514d16ccf806818ce91ab77937365af90bbf74a35be6b40b8eedf2785e42
-            874d");
-        assert_eq!(&expected[..], &buf[64..]);
-    }
-
-    #[test]
-    fn xchacha20_case_1() {
-        let key = hex!("82f411a074f656c66e7dbddb0a2c1b22760b9b2105f4ffdbb1d4b1e824e21def");
-        let nonce = hex!("3b07ca6e729eb44a510b7a1be51847838a804f8b106b38bd");
-        let mut state = XChaCha20::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-        let mut xs = [0u8; 100];
-        state.apply_keystream(&mut xs);
-        let expected = hex!("
-            201863970b8e081f4122addfdf32f6c03e48d9bc4e34a59654f49248b9be59d3eaa106ac3376e7e7d9d1251f2cbf61ef27000f3d19afb76b
-            9c247151e7bc26467583f520518eccd2055ccd6cc8a195953d82a10c2065916778db35da2be44415d2f5efb0");
-        assert_eq!(&expected[..], &xs[..]);
-    }
-
-    #[test]
-    fn seek_off_end() {
-        let mut st = Ietf::new(
-            GenericArray::from_slice(&[0xff; 32]),
-            GenericArray::from_slice(&[0; 12]),
-        );
-        st.seek(0x40_0000_0000);
-
-        assert!(st.try_apply_keystream(&mut [0u8; 1]).is_err());
-    }
-
-    #[test]
-    fn read_last_bytes() {
-        let mut st = Ietf::new(
-            GenericArray::from_slice(&[0xff; 32]),
-            GenericArray::from_slice(&[0; 12]),
-        );
-
-        st.seek(0x40_0000_0000 - 10);
-        st.apply_keystream(&mut [0u8; 10]);
-        assert!(st.try_apply_keystream(&mut [0u8; 1]).is_err());
-
-        st.seek(0x40_0000_0000 - 10);
-        assert!(st.try_apply_keystream(&mut [0u8; 11]).is_err());
-    }
-
-    #[test]
-    fn seek_consistency() {
-        let mut st = Ietf::new(
-            GenericArray::from_slice(&[50; 32]),
-            GenericArray::from_slice(&[44; 12]),
-        );
-
-        let mut continuous = [0u8; 1000];
-        st.apply_keystream(&mut continuous);
-
-        let mut chunks = [0u8; 1000];
-
-        st.seek(128);
-        st.apply_keystream(&mut chunks[128..300]);
-
-        st.seek(0);
-        st.apply_keystream(&mut chunks[0..10]);
-
-        st.seek(300);
-        st.apply_keystream(&mut chunks[300..533]);
-
-        st.seek(533);
-        st.apply_keystream(&mut chunks[533..]);
-
-        st.seek(10);
-        st.apply_keystream(&mut chunks[10..128]);
-
-        assert_eq!(&continuous[..], &chunks[..]);
-    }
-
-    #[test]
-    fn wide_matches_narrow() {
-        let key = hex!("fa44478c59ca70538e3549096ce8b523232c50d9e8e8d10c203ef6c8d07098a5");
-        let nonce = hex!("8d3a0d6d7827c007");
-        let mut buf = [0; 2048];
-        let mut state = ChaCha20::new(
-            GenericArray::from_slice(&key),
-            GenericArray::from_slice(&nonce),
-        );
-
-        let lens = [
-            2048, 2047, 1537, 1536, 1535, 1025, 1024, 1023, 768, 513, 512, 511, 200, 100, 50,
-        ];
-
-        for &len in &lens {
-            let buf = &mut buf[0..len];
-
-            // encrypt with hybrid wide/narrow
-            state.seek(0);
-            state.apply_keystream(buf);
-            state.seek(0);
-            // decrypt with narrow only
-            state.try_apply_keystream_narrow(buf).unwrap();
-            for &byte in buf.iter() {
-                assert_eq!(byte, 0);
-            }
-
-            // encrypt with hybrid wide/narrow
-            let offset = 0x3fffffff70u64;
-            state.seek(offset);
-            state.apply_keystream(buf);
-            // decrypt with narrow only
-            state.seek(offset);
-            state.try_apply_keystream_narrow(buf).unwrap();
-            for &byte in buf.iter() {
-                assert_eq!(byte, 0);
-            }
-        }
-    }
-}
+#[cfg(feature = "rustcrypto_api")]
+mod rustcrypto_impl;
+#[cfg(feature = "rustcrypto_api")]
+pub use self::rustcrypto_impl::{stream_cipher, ChaCha12, ChaCha20, ChaCha8, Ietf, XChaCha20};
