+++ utils-simd/ppv-lite86/Cargo.toml
@@ -10,7 +10,6 @@ keywords = ["crypto", "simd", "x86"]
 categories = ["cryptography", "no-std"]
 
 [dependencies]
-crypto-simd = "0.1"
 
 [badges]
 travis-ci = { repository = "cryptocorrosion/cryptocorrosion" }
