+++ utils-simd/ppv-lite86/src/x86_64/avx.rs
